Remove unneeded bff token handler
